Move test access helper into test package
